requirements.txt
Flask
Flask-Login
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
psycopg2
requests